Added firebase functions to project
currently have a single function working which takes a request from the front end (Payment.js) using Stripe's API and triggers a payment request on Stripe.

Working to get functionality set up to store the order in firestore db as well so orders can be retrieved and displayed on the orders page (once it is built)

Already configured env variables for the firebase functions, both for the local emulator and the deployment environment.